move gui to component
