+++ src/Client.css
@@ -1,0 +1,3 @@
+table {
+    background-color: #222222;
+}
